requirements.txt
bcrypt==4.1.3
certifi==2024.7.4
charset-normalizer==3.3.2
faiss-cpu==1.8.0.post1
greenlet==3.0.3
icalendar==5.0.13
idna==3.7
joblib==1.4.2
marshmallow==3.21.3
numpy==1.26.4
packaging==24.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
SQLAlchemy==2.0.31
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.2
